# api/openapi.json
{
  "components": {
    "schemas": {
      "AcceptanceItem": {
        "description": "Spec §13.3. One sentence someone can check, numbered within its task.",
        "properties": {
          "number": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "number",
          "text"
        ],
        "type": "object"
      },
...
      "ApprovePlan": {
        "properties": {
          "command_id": {
            "description": "The idempotency key (spec §13.5), scoped to the plan version.",
            "type": "string"
          },
          "expected_revision": {
            "description": "The revision the person saw; approval of any other is refused.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "command_id",
          "expected_revision"
        ],
        "type": "object"
      },
      "Approved": {
        "description": "What an approval did, fixed when it committed (§13.2, §13.5).",
        "properties": {
          "frozen_at": {
            "type": "string"
          },
          "revision": {
            "format": "int64",
            "type": "integer"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          },
          "workflow_id": {
            "$ref": "#/components/schemas/WorkflowId"
          }
        },
        "required": [
          "workflow_id",
          "version",
          "revision",
          "frozen_at"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "The body of every error this API answers: the stable code a client\nmatches on (spec §3.4), and a human message nobody should match on. A\nrequest axum rejects before a handler runs (malformed JSON, a missing query\nparameter) gets one too, through\n[`rejections_as_error_bodies`].",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "current_revision": {
            "description": "`REVISION_CONFLICT` only: the resource's revision now, to read again at.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "problems": {
            "description": "`WORKFLOW_VALIDATION_FAILED` only: each problem, the same sentences\n`message` joins.",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ErrorCode": {
        "description": "Stable codes clients pattern-match on. Never match on human text.\nSpec §3.4. `Blocked`/`Rejected` are domain outcomes and never appear here.",
        "enum": [
          "HARNESS_START_FAILED",
          "PROCESS_SPAWN_FAILED",
          "PROCESS_TERMINATED",
          "PROCESS_TERMINATION_FAILED",
          "RUNTIME_STOPPING",
          "STORAGE_UNAVAILABLE",
          "STORAGE_MIGRATION_FAILED",
          "STORAGE_CONSTRAINT_VIOLATION",
          "COMMAND_CONFLICT",
          "IDEMPOTENCY_KEY_REQUIRED",
          "INVALID_COMMAND",
          "INVALID_CURSOR",
          "AGENT_AUTH_FAILED",
          "AGENT_UNSUPPORTED_PROFILE",
          "PATH_INVALID",
          "PATH_NOT_FOUND",
          "PATH_NOT_A_DIRECTORY",
          "PATH_ACCESS_DENIED",
          "PATH_ALREADY_EXISTS",
          "PATH_UNAVAILABLE",
          "ORIGIN_REFUSED",
          "HARNESS_UNAVAILABLE",
          "SETTING_NOT_OFFERED",
          "MODE_NOT_ALLOWED",
          "HARNESS_LOCKED",
          "THREAD_BUSY",
          "FORK_POINT_NOT_SUPPORTED",
          "WORKFLOW_FROZEN_IMMUTABLE",
          "WORKFLOW_VALIDATION_FAILED",
          "REVISION_CONFLICT",
          "PROJECT_HAS_THREADS",
          "GRANT_SCOPE",
          "GRANT_INVALID"
        ],
        "type": "string"
      },
...
      "GrantId": {
        "format": "uuid",
        "type": "string"
      },
...
      "LastEdit": {
        "description": "The edit that set a version's current revision (§13.10's plan read).",
        "properties": {
          "changed_tasks": {
            "items": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            },
            "type": "array"
          },
          "revision": {
            "format": "int64",
            "type": "integer"
          },
          "summary": {
            "type": "string"
          }
        },
        "required": [
          "revision",
          "summary",
          "changed_tasks"
        ],
        "type": "object"
      },
...
      "Link": {
        "description": "Spec §13.3. `task` waits for `after`.",
        "properties": {
          "after": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/LinkKind"
          },
          "label": {
            "description": "A few words saying what passes from `after` to `task`.",
            "type": "string"
          },
          "task": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "waiting_items": {
            "description": "The acceptance items of `task` that wait. Empty for `needs`.",
            "items": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            },
            "type": "array"
          }
        },
        "required": [
          "task",
          "after",
          "kind",
          "label"
        ],
        "type": "object"
      },
      "LinkKind": {
        "description": "Spec §13.3. How a link's `task` waits for its `after`.",
        "enum": [
          "needs",
          "completes_after"
        ],
        "type": "string"
      },
...
      "Plan": {
        "description": "One stored version of a plan, as a reader sees it now (§13.2).",
        "properties": {
          "blockers": {
            "description": "What blocks approval ([`approval_problems`]) for a `Draft`; empty for\na `Frozen` version.",
            "items": {
              "$ref": "#/components/schemas/Problem"
            },
            "type": "array"
          },
          "change_reason": {
            "description": "Why this version was started (§16.3): `None` for v1, and for a version\nfrom before migration 0012, which reads \"Reason not recorded\".",
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "frozen_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "goal": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/WorkflowId"
          },
          "last_edit": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LastEdit",
                "description": "`None` until the version is first edited."
              }
            ]
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "array"
          },
          "next": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WorkflowId",
                "description": "The version copied from this one, once it exists."
              }
            ]
          },
          "plan_id": {
            "$ref": "#/components/schemas/PlanId"
          },
          "plan_state": {
            "$ref": "#/components/schemas/PlanState"
          },
          "previous": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WorkflowId",
                "description": "The version this one was copied from."
              }
            ]
          },
          "project_id": {
            "$ref": "#/components/schemas/ProjectId"
          },
          "revision": {
            "format": "int64",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/WorkflowState"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/PlanTask"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          },
          "written_by": {
            "$ref": "#/components/schemas/WrittenBy",
            "description": "Who wrote this version (§16.3)."
          }
        },
        "required": [
          "id",
          "plan_id",
          "plan_state",
          "project_id",
          "written_by",
          "version",
          "revision",
          "state",
          "title",
          "goal",
          "tasks",
          "links",
          "blockers",
          "created_at"
        ],
        "type": "object"
      },
      "PlanCommand": {
        "properties": {
          "command_id": {
            "description": "The idempotency key (spec §16.2), scoped to the plan.",
            "type": "string"
          }
        },
        "required": [
          "command_id"
        ],
        "type": "object"
      },
      "PlanId": {
        "format": "uuid",
        "type": "string"
      },
      "PlanListing": {
        "description": "A plan as a project's list shows it: its latest version (`id`).",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/WorkflowId"
          },
          "plan_id": {
            "$ref": "#/components/schemas/PlanId"
          },
          "plan_state": {
            "$ref": "#/components/schemas/PlanState"
          },
          "state": {
            "$ref": "#/components/schemas/WorkflowState"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "plan_id",
          "plan_state",
          "id",
          "title",
          "version",
          "state",
          "updated_at"
        ],
        "type": "object"
      },
      "PlanState": {
        "description": "Spec §16.2. An `Archived` plan is read, never written.",
        "enum": [
          "Active",
          "Archived"
        ],
        "type": "string"
      },
      "PlanTask": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TaskContent"
          },
          {
            "properties": {
              "id": {
                "$ref": "#/components/schemas/TaskId"
              }
            },
            "required": [
              "id"
            ],
            "type": "object"
          }
        ],
        "description": "One task of a stored version: its storage id beside its content."
      },
      "PlanVersions": {
        "description": "One plan with every version, oldest first (§16.10's GET /api/plans/{id}).",
        "properties": {
          "archived_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "plan_id": {
            "$ref": "#/components/schemas/PlanId"
          },
          "project_id": {
            "$ref": "#/components/schemas/ProjectId"
          },
          "state": {
            "$ref": "#/components/schemas/PlanState"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionLine"
            },
            "type": "array"
          }
        },
        "required": [
          "plan_id",
          "project_id",
          "state",
          "versions"
        ],
        "type": "object"
      },
...
      "Problem": {
        "description": "One broken or missing thing, named so a person can act on it: tasks as `T4`.",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
      "ProjectId": {
        "format": "uuid",
        "type": "string"
      },
...
      "TaskContent": {
        "description": "Spec §13.3. A task as the plan holds it, shown as `T{number}`.",
        "properties": {
          "acceptance": {
            "items": {
              "$ref": "#/components/schemas/AcceptanceItem"
            },
            "type": "array"
          },
          "goal": {
            "type": "string"
          },
          "number": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "reads": {
            "description": "Paths the task reads (its declared scope).",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "writes": {
            "description": "Paths the task may change (its declared scope).",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "number",
          "title",
          "goal",
          "reads",
          "writes",
          "acceptance"
        ],
        "type": "object"
      },
      "TaskId": {
        "format": "uuid",
        "type": "string"
      },
...
      "ThreadId": {
        "format": "uuid",
        "type": "string"
      },
...
      "VersionLine": {
        "properties": {
          "change_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/WorkflowState"
          },
          "title": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          },
          "workflow_id": {
            "$ref": "#/components/schemas/WorkflowId"
          },
          "written_by": {
            "$ref": "#/components/schemas/WrittenBy"
          }
        },
        "required": [
          "workflow_id",
          "version",
          "state",
          "title",
          "written_by",
          "created_at"
        ],
        "type": "object"
      },
...
      "WorkflowId": {
        "format": "uuid",
        "type": "string"
      },
      "WorkflowState": {
        "description": "Spec §13.2. A `Draft` is edited; a `Frozen` version is approved and never\nchanges again. The web client shows `Frozen` as \"Approved\".",
        "enum": [
          "Draft",
          "Frozen"
        ],
        "type": "string"
      },
      "WrittenBy": {
        "description": "Spec §16.3: who wrote a version, recorded once when it was created.",
        "oneOf": [
          {
            "description": "The internal Planner: its conversation, and the turn when one was\nrecorded (none before migration 0012).",
            "properties": {
              "harness": {
                "description": "`agent_invocation.harness_kind`, e.g. `claude-code`.",
                "type": [
                  "string",
                  "null"
                ]
              },
              "kind": {
                "enum": [
                  "planner"
                ],
                "type": "string"
              },
              "model": {
                "description": "The observed model, else the requested one; `None` without a turn.",
                "type": [
                  "string",
                  "null"
                ]
              },
              "thread_id": {
                "$ref": "#/components/schemas/ThreadId"
              },
              "thread_removed": {
                "type": "boolean"
              },
              "thread_title": {
                "type": "string"
              }
            },
            "required": [
              "thread_id",
              "thread_title",
              "thread_removed",
              "kind"
            ],
            "type": "object"
          },
          {
            "description": "An external agent's project grant.",
            "properties": {
              "grant_id": {
                "$ref": "#/components/schemas/GrantId"
              },
              "kind": {
                "enum": [
                  "external"
                ],
                "type": "string"
              }
            },
            "required": [
              "grant_id",
              "kind"
            ],
            "type": "object"
          }
        ]
      },
...
    }
  },
  "info": {
    "description": "The local daemon's HTTP API. Every client is on another origin; only origins passed with `--allow-origin` may call it. There is no authentication: the daemon binds to loopback (spec §1, OPEN block on remote access). Errors answered by the API carry `ErrorBody`, whose `code` is stable; a request axum rejects before a handler runs (malformed JSON, a missing parameter) is answered in plain text.",
    "license": {
      "name": ""
    },
    "title": "shadows",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/plans/{id}": {
      "get": {
        "operationId": "get_plan_versions",
        "parameters": [
          {
            "description": "The plan",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PlanId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanVersions"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "INVALID_COMMAND: no such plan"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE"
          }
        },
        "summary": "One plan with every version, oldest first (§16.10).",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/plans/{id}/archive": {
      "post": {
        "operationId": "archive_plan",
        "parameters": [
          {
            "description": "The plan",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PlanId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanVersions"
                }
              }
            },
            "description": "Archived, or the replay of the same command"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "INVALID_COMMAND: no such plan"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "COMMAND_CONFLICT"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE"
          }
        },
        "summary": "Archives a plan (§16.2). An archived plan is read, never written.\nArchiving an already archived plan answers the plan unchanged.",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/plans/{id}/unarchive": {
      "post": {
        "operationId": "unarchive_plan",
        "parameters": [
          {
            "description": "The plan",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PlanId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanVersions"
                }
              }
            },
            "description": "Unarchived, or the replay of the same command"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "INVALID_COMMAND: no such plan"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "COMMAND_CONFLICT"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE"
          }
        },
        "summary": "Unarchives a plan (§16.2), returning it to `Active`.\nUnarchiving an active plan answers the plan unchanged.",
        "tags": [
          "workflows"
        ]
      }
    },
...
    "/api/projects/{id}/workflows": {
      "get": {
        "operationId": "list_plans",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          },
          {
            "description": "When true, includes archived plans in the listing (§16.2).",
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanListing"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE"
          }
        },
        "summary": "The project's plans, each by its latest version, in the order\nthey were created. Active plans only, or archived ones too when `archived=true`.\nAn unknown project has none.",
        "tags": [
          "workflows"
        ]
      }
    },
...
    "/api/workflows/{id}": {
      "get": {
        "operationId": "get_plan",
        "parameters": [
          {
            "description": "The plan version",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkflowId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "INVALID_COMMAND: no such plan version"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE"
          }
        },
        "summary": "One plan version: its tasks, links and revision, the versions before and\nafter it, what blocks its approval, and what the last edit changed.",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{id}/approve": {
      "post": {
        "operationId": "approve_plan",
        "parameters": [
          {
            "description": "The plan version",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkflowId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovePlan"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Approved"
                }
              }
            },
            "description": "Approved, or the replay of the same command"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "INVALID_COMMAND: no such plan version"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "REVISION_CONFLICT, carrying `current_revision`; WORKFLOW_FROZEN_IMMUTABLE; COMMAND_CONFLICT"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "WORKFLOW_VALIDATION_FAILED, carrying `problems`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE"
          }
        },
        "summary": "Approves a `Draft` version as the person saw it: it becomes `Frozen` and\nnever changes again (§13.2). The client reads the plan again afterwards. A\nreplay of the same command answers what it answered first.",
        "tags": [
          "workflows"
        ]
      }
    }
  },
...
}